[plasma] move tests to test_thermo.py
Use properties such as standard_enthalpies_RT
